eman

+++ checking.h
@@ -12,7 +12,7 @@ class CheckingAccount: public Account
     double overDraftLimit = 100.00;
     public:
     CheckingAccount(string _fName, string _lName, string _address, string _phone, string _email, int _id, double _balance, double _overDraftLimit);
-    void Withdraw(int amount);
+    void Withdraw(double amount);
 
 
 };
